Move Spigot implementation module to Spigot module. Add NPCv1_8_R3 implementation. Add spigot main classes. Refractoring.
