+++ elitenpcs-api/src/main/java/io/github/elitedev/api/NPCPlugin.java
@@ -1,0 +1,9 @@
+package io.github.elitedev.api;
+
+import io.github.elitedev.api.io.Storage;
+
+public interface NPCPlugin {
+
+    Storage getStorage();
+
+}
