+++ elitenpcs-api/src/main/java/io/github/elitedev/api/exceptions/InvalidNpcException.java
@@ -1,0 +1,8 @@
+package io.github.elitedev.api.exceptions;
+
+public class InvalidNpcException extends Exception {
+
+    public InvalidNpcException() {
+        super("NPC or NPC UUID cannot be null");
+    }
+}
