+++ elitenpcs-api/src/main/java/io/github/elitedev/api/io/StorageOperationMode.java
@@ -1,0 +1,8 @@
+package io.github.elitedev.api.io;
+
+public enum StorageOperationMode {
+
+    IMMEDIATE,
+    ASYNC
+
+}
