--- elitenpcs-api/src/main/java/io/github/elitedev/exceptions/InvalidNpcException.java
@@ -1,10 +1,0 @@
-package io.github.elitedev.exceptions;
-
-import io.github.elitedev.entity.NPC;
-
-public class InvalidNpcException extends Exception {
-
-    public InvalidNpcException() {
-        super("NPC or NPC UUID cannot be null");
-    }
-}
