--- elitenpcs-api/src/main/java/io/github/elitedev/io/StorageOperationMode.java
@@ -1,8 +1,0 @@
-package io.github.elitedev.io;
-
-public enum StorageOperationMode {
-
-    IMMEDIATE,
-    ASYNC
-
-}
