--- elitenpcs-spigot/src/main/java/io/github/elitedev/entity/SpigotNPCRegistry.java
@@ -1,8 +1,0 @@
-package io.github.elitedev.entity;
-
-public interface SpigotNPCRegistry extends NPCRegistry {
-
-    static SpigotNPCRegistry DEFAULT_IMPL () {
-        return new SpigotNPCRegistryImpl();
-    }
-}
