+++ elitenpcs-spigot/src/main/java/io/github/elitedev/spigot/entity/SpigotNPCRegistry.java
@@ -1,0 +1,10 @@
+package io.github.elitedev.spigot.entity;
+
+import io.github.elitedev.api.entity.NPCRegistry;
+
+public interface SpigotNPCRegistry extends NPCRegistry {
+
+    static SpigotNPCRegistry DEFAULT_IMPL () {
+        return new SpigotNPCRegistryImpl();
+    }
+}
