+++ elitenpcs-spigot/src/main/java/io/github/elitedev/spigot/event/AbstractEvent.java
@@ -1,0 +1,18 @@
+package io.github.elitedev.spigot.event;
+
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+
+abstract class AbstractEvent extends Event {
+
+    private HandlerList handlerList;
+
+    AbstractEvent() {
+        handlerList = new HandlerList();
+    }
+
+    @Override
+    public HandlerList getHandlers() {
+        return handlerList;
+    }
+}
