+++ pom.xml
@@ -12,7 +12,6 @@
     <modules>
         <module>elitenpcs-api</module>
         <module>elitenpcs-spigot</module>
-        <module>v1_8_R3</module>
     </modules>
 
     <build>
@@ -25,6 +24,20 @@
                     <source>1.8</source>
                     <target>1.8</target>
                 </configuration>
+            </plugin>
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+                <version>3.2.1</version>
+                <executions>
+                    <execution>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
